moving all lookup files into update_db folder adding return key binding for search
